requirements.txt
django
djangorestframework
django-rest-swagger
drf-nested-routers
docopts
